Change int/Integer to double/Double to handle needed precision

+++ PracticalProjects/src/io/github/nathanvegetable/numbers/FindPiToTheNthDigit.java
@@ -5,10 +5,10 @@ public class FindPiToTheNthDigit {
 		FindPiToTheNthDigit.findWithLimit(10);
 	}
 
-	public static int findWithLimit(int limit) {
+	public static double findWithLimit(int limit) {
 		double pi = Math.PI;
 		String piString = String.valueOf(pi).replace(".", "");
 		piString = piString.substring(0, limit + 1);
-		return Integer.parseInt(piString);
+		return Double.parseDouble(piString);
 	}
 }
